chore(cli): remove bundled UI, add lightweight env catalog, quiet serve placeholder
