--- world_models/ui/__init__.py
@@ -1,3 +1,0 @@
-from .server import app
-
-__all__ = ["app"]
